+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 3.0.0 [unreleased]
 
+- Detect numeric axes automatically (and removed `xtype` option)
 - Removed `window.Chartkick = {...}` way to set config
 - Normalize bubble sizes